+++ app/utils/utils.js
@@ -1,50 +1,53 @@
 const i18next = require('i18next')
-const moment = require('moment')
-moment().format()
 
-const formatRemaining = function (seconds) {
+// Returns `XXX remaining`
+// `About 2 hours 53 mintes remaining`
+// `About 53 mintes remaining`
+// `10 seconds remaining`
+const formatTimeRemaining = function (milliseconds) {
+  const seconds = milliseconds / 1000
   if (seconds < 60) {
-    return i18next.t('utils.secondsLeft', { count: seconds + 1 })
+    return i18next.t('utils.secondsRemaining', { seconds: Math.trunc(seconds + 1) })
   }
-  return i18next.t('utils.minutesLeft', { count: Math.trunc((seconds / 60) + 1) })
+  if (seconds >= 60 && seconds < 3600) {
+    return i18next.t('utils.aboutMinutesRemaining', { minutes: Math.trunc((seconds / 60) + 1) })
+  }
+  if (seconds >= 3600) {
+    const hours = Math.trunc(seconds / 3600)
+    const minutes = Math.trunc(((seconds - hours * 3600) / 60) + 1)
+    if (seconds % 3600 === 0) {
+      return i18next.t('utils.aboutHoursRemaining', { hours: hours })
+    } else if (minutes === 60) {
+      return i18next.t('utils.aboutHoursRemaining', { hours: hours + 1 })
+    } else {
+      return i18next.t('utils.aboutHoursMinutesRemaining', { minutes: minutes, hours: hours })
+    }
+  }
 }
 
-const formatTillBreak = function (milliseconds) {
-  const minutes = Math.round(milliseconds / 60000)
-  if (minutes < 1) {
-    const seconds = Math.round((milliseconds % 60000) / 5000) * 5
-    return i18next.t('utils.s', { seconds: seconds })
+// Returns `XXX remaining`
+// `in about 2 hours 53 mintes`
+// `in about 53 mintes`
+// `in 10 seconds`
+const formatTimeIn = function (milliseconds) {
+  const seconds = milliseconds / 1000
+  if (seconds < 60) {
+    return i18next.t('utils.inSeconds', { seconds: Math.trunc(seconds + 1) })
   }
-  return i18next.t('utils.m', { minutes: minutes })
-}
-
-const formatPauseTimeLeft = function (milliseconds) {
-  let timeString = ''
-  let hours = Math.floor(milliseconds / (1000 * 3600))
-  const remainder = (milliseconds - hours * 1000 * 3600)
-  let minutes = Math.floor(remainder / 60000)
-  if (minutes >= 60) {
-    minutes -= 60
-    hours += 1
+  if (seconds >= 60 && seconds < 3600) {
+    return i18next.t('utils.inAboutMinutes', { minutes: Math.trunc((seconds / 60) + 1) })
   }
-  if (hours >= 1) {
-    timeString += i18next.t('utils.h', { hours: hours })
+  if (seconds >= 3600) {
+    const hours = Math.trunc(seconds / 3600)
+    const minutes = Math.trunc(((seconds - hours * 3600) / 60) + 1)
+    if (seconds % 3600 === 0) {
+      return i18next.t('utils.inAboutHours', { hours: hours })
+    } else if (minutes === 60) {
+      return i18next.t('utils.inAboutHours', { hours: hours + 1 })
+    } else {
+      return i18next.t('utils.inAboutHoursMinutes', { minutes: minutes, hours: hours })
+    }
   }
-  if (minutes >= 1) {
-    timeString += i18next.t('utils.m', { minutes: minutes })
-  }
-  if (minutes < 1 && hours < 1) {
-    timeString = `${i18next.t('utils.lessThan1m')}`
-  }
-  return timeString
-}
-
-function formatTimeOfNextBreak (timeLeft) {
-  const date = moment.utc(Date.now() + timeLeft).local()
-  const hours = String(date.hours())
-  const minutes = String(date.minutes()).padStart(2, '0')
-
-  return [hours, minutes]
 }
 
 // does not consider `postponesLimit`
@@ -58,10 +61,8 @@ function canSkip (strictMode, postpone, passedPercent, postponePercent) {
 }
 
 module.exports = {
-  formatRemaining,
-  formatTillBreak,
-  formatPauseTimeLeft,
-  formatTimeOfNextBreak,
+  formatTimeRemaining,
+  formatTimeIn,
   canPostpone,
   canSkip
 }